cleanup: Do some renaming and add godoc.
